requirements.txt
flask
flask_mail
python-dotenv